Follow style guidelines, refactoring
Can't click "Save & continue" when there are inputs that are blank
